feat(movement): elf movement improvements

+++ src/map/system.rs
@@ -2,8 +2,8 @@ use crate::map::Map;
 use bevy::prelude::Commands;
 
 pub fn setup_map(mut commands: Commands) {
-    const MAP_WIDTH: f32 = 16.0;
-    const MAP_HEIGHT: f32 = 9.0;
+    const MAP_WIDTH: u8 = 16;
+    const MAP_HEIGHT: u8 = 9;
 
     commands.insert_resource(Map {
         width: MAP_WIDTH,
